+++ src/components/util/Icon.tsx
@@ -15,20 +15,31 @@ export enum IconType {
   PanelIcon = 'panel-icon'
 }
 
+export enum IconContainerSize {
+  Small = 'is-small',
+  Normal = '',
+  Medium = 'is-medium',
+  Large = 'is-large'
+}
+
 export interface IconProps {
   name: string;
-  style: IconStyle;
+  style?: IconStyle;
   size?: IconSize;
   type?: IconType;
+  containerSize?: IconContainerSize;
+  left?: boolean;
 }
 
 export const Icon = ({
   name,
-  style,
+  style = IconStyle.Solid,
   size = IconSize.Normal,
-  type = IconType.Icon
+  type = IconType.Icon,
+  containerSize = IconContainerSize.Normal,
+  left = false
 }: IconProps) => (
-  <span className={type}>
+  <span className={`${type} ${containerSize} ${left ? 'is-left' : ''}`}>
     <i className={`${style} fa-${name} ${size}`}></i>
   </span>
 );